+++ django_resized/tests/tests.py
@@ -14,7 +14,7 @@ from .models import Product
 class ResizeTest(TestCase):
 
     def tearDown(self):
-        dirname = os.path.join(settings.MEDIA_ROOT, 'testapp')
+        dirname = os.path.join(settings.MEDIA_ROOT, 'tests')
         if os.path.exists(dirname):
             shutil.rmtree(dirname)
 
@@ -23,28 +23,28 @@ class ResizeTest(TestCase):
             image1=File(open('media/big.jpg', 'rb')),
         )
         im1 = Image.open(product.image1.path)
-        self.assertEquals(im1.size, (467, 350))
+        self.assertEqual(im1.size, (467, 350))
 
     def test_resizes_settings_default(self):
         product = Product.objects.create(
             image2=File(open('media/big.jpg', 'rb')),
         )
         im2 = Image.open(product.image2.path)
-        self.assertEquals(im2.size, (400, 300))
+        self.assertEqual(im2.size, (400, 300))
 
     def test_resize_crop_center(self):
         product = Product.objects.create(
             image3=File(open('media/big.jpg', 'rb')),
         )
         im3 = Image.open(product.image3.path)
-        self.assertEquals(im3.size, (40, 40))
+        self.assertEqual(im3.size, (40, 40))
 
     def test_resize_crop_right(self):
         product = Product.objects.create(
             image4=File(open('media/big.jpg', 'rb')),
         )
         im4 = Image.open(product.image4.path)
-        self.assertEquals(im4.size, (100, 100))
+        self.assertEqual(im4.size, (100, 100))
 
     def test_resize_with_quality(self):
         product = Product.objects.create(
@@ -73,7 +73,7 @@ class ResizeTest(TestCase):
             image7=File(open('media/big.jpg', 'rb')),
         )
         im7 = Image.open(product.image7.path)
-        self.assertEquals(im7.size, (604, 453))
+        self.assertEqual(im7.size, (604, 453))
 
     def test_force_format(self):
         product = Product.objects.create(